requirements.txt
# ASR-RAG (Python 3.10–3.12). Install: pip install -r requirements.txt
# Then: python -m spacy download en_core_web_sm
# Optional CHEMICAL/DISEASE F1: pip install scispacy + en_ner_bc5cdr_md (see README)

datasets>=2.14.0
soundfile>=0.12.0
numpy>=1.24.0,<2.0.0
textgrid>=1.6.0

openai-whisper>=20231117
ollama>=0.3.0

rank_bm25>=0.2.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

jiwer>=3.0.0
bert-score>=0.3.13
rouge-score>=0.1.2
nltk>=3.8.0

spacy>=3.5.0,<4.0.0

matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=1.5.0
